requirements.txt
# Framework dan Server
fastapi
uvicorn[standard]

# Database dan ORM
sqlalchemy
aiomysql
greenlet  # <-- Diperlukan untuk SQLAlchemy async

# Upload File dan Request HTTP
python-multipart
httpx

# Pemrosesan Gambar
pillow
opencv-python-headless

# Konfigurasi
python-dotenv
cryptography
boto3